Renamed project from 'Prime Number Calculator' to 'LCM Calculator'
