+++ src/main/java/com/west/assignment/least_common_multiple_calculator/LeastCommonMultipleCalculatorApplication.java
@@ -1,0 +1,13 @@
+package com.west.assignment.least_common_multiple_calculator;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LeastCommonMultipleCalculatorApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(LeastCommonMultipleCalculatorApplication.class, args);
+    }
+
+}
