--- src/main/java/com/west/assignment/prime_number_calculator/PrimeNumberCalculatorApplication.java
@@ -1,13 +1,0 @@
-package com.west.assignment.prime_number_calculator;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PrimeNumberCalculatorApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(PrimeNumberCalculatorApplication.class, args);
-    }
-
-}
