+++ src/test/java/com/west/assignment/least_common_multiple_calculator/LeastCommonMultipleCalculatorApplicationTests.java
@@ -1,10 +1,10 @@
-package com.west.assignment.prime_number_calculator;
+package com.west.assignment.least_common_multiple_calculator;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class PrimeNumberCalculatorApplicationTests {
+class LeastCommonMultipleCalculatorApplicationTests {
 
     @Test
     void contextLoads() {
